+++ scripts/screens/ChooseAdoptiveParentScreen.py
@@ -8,9 +8,9 @@ from scripts.cat.cats import Cat
 from scripts.game_structure import image_cache
 from scripts.game_structure.propagating_thread import PropagatingThread
 from scripts.game_structure.ui_elements import (
-    UIImageButton,
     UISpriteButton,
 )
+from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
 from ..ui.scale import ui_scale, ui_scale_dimensions, ui_scale_offset
