+++ scripts/screens/ClanSettingsScreen.py
@@ -12,10 +12,10 @@ import ujson
 from scripts.cat.cats import Cat
 from scripts.game_structure.game.settings import game_setting_get
 from scripts.game_structure.ui_elements import (
-    UIImageButton,
     UIModifiedScrollingContainer,
     UICheckbox,
 )
+from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.housekeeping.datadir import open_data_dir
 from ..ui.theme import get_text_box_theme
