+++ scripts/screens/LeaderDenScreen.py
@@ -16,9 +16,9 @@ from scripts.clan_package.settings.clan_settings import (
 from scripts.game_structure import constants
 from scripts.game_structure.screen_settings import MANAGER
 from scripts.game_structure.ui_elements import (
-    UIImageButton,
     UISpriteButton,
 )
+from scripts.ui.elements.image_button import UIImageButton
 from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.screens.Screens import Screens
 from scripts.ui.generate_box import get_box, BoxStyles
