+++ scripts/screens/ListScreen.py
@@ -21,10 +21,10 @@ from scripts.cat.enums import CatGroup
 from scripts.game_structure import game
 from scripts.game_structure.screen_settings import game_screen_size, MANAGER
 from scripts.game_structure.ui_elements import (
-    UIImageButton,
     UICatListDisplay,
     UIDropDown,
 )
+from scripts.ui.elements.image_button import UIImageButton
 from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.screens.Screens import Screens
 from scripts.screens.enums import GameScreen
