+++ scripts/screens/MakeClanScreen.py
@@ -16,9 +16,9 @@ from scripts.events_module.patrol.patrol import Patrol
 from scripts.game_structure import image_cache, constants
 from scripts.game_structure import game
 from scripts.game_structure.ui_elements import (
-    UIImageButton,
     UISpriteButton,
 )
+from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
 from ..ui.scale import ui_scale, ui_scale_dimensions, ui_scale_offset, ui_scale_blit
