+++ scripts/screens/MedDenScreen.py
@@ -9,10 +9,10 @@ from scripts.clan_resources.herb.herb_supply import MESSAGES
 from scripts.game_structure import game
 from scripts.game_structure.ui_elements import (
     UISpriteButton,
-    UIImageButton,
     UITextBoxTweaked,
     UIModifiedImage,
 )
+from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import event_text_adjust, shorten_text_to_fit
