+++ scripts/screens/MediationScreen.py
@@ -8,10 +8,10 @@ import pygame_gui.elements
 from scripts.cat.cats import Cat
 from scripts.game_structure import image_cache, game
 from scripts.game_structure.ui_elements import (
-    UIImageButton,
     UISpriteButton,
     UIRelationDisplay,
 )
+from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import shorten_text_to_fit
