+++ scripts/screens/RelationshipScreen.py
@@ -8,10 +8,10 @@ from scripts.cat.cats import Cat
 from scripts.game_structure import image_cache, constants
 from scripts.game_structure import game
 from scripts.game_structure.ui_elements import (
-    UIImageButton,
     UISpriteButton,
     UIRelationDisplay,
 )
+from scripts.ui.elements.image_button import UIImageButton
 from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.ui.windows.relationship_log import RelationshipLogWindow
 from scripts.screens.Screens import Screens
