+++ scripts/screens/SpriteInspectScreen.py
@@ -6,7 +6,7 @@ import pygame_gui
 
 from scripts.cat.cats import Cat
 from scripts.game_structure import game
-from scripts.game_structure.ui_elements import UIImageButton
+from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import shorten_text_to_fit
