+++ scripts/screens/SwitchClanScreen.py
@@ -9,7 +9,7 @@ from pygame_gui.elements import UIImage
 import scripts.game_structure.screen_settings
 from scripts.clan import Clan
 from scripts.game_structure import game
-from scripts.game_structure.ui_elements import UIImageButton
+from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.ui.windows.delete_check import CheckDeletionWindow
 from ..ui.theme import get_text_box_theme
