+++ scripts/ui/windows/change_cat_toggles.py
@@ -2,9 +2,7 @@ import pygame
 import pygame_gui
 
 from scripts.game_structure import game
-from scripts.game_structure.ui_elements import (
-    UIImageButton,
-)
+from scripts.ui.elements.image_button import UIImageButton
 from scripts.screens.enums import GameScreen
 from scripts.ui.windows.window_base_class import GameWindow
 from scripts.ui.scale import ui_scale
