+++ scripts/ui/windows/herbs.py
@@ -15,9 +15,9 @@ from scripts.game_structure.game.switches import (
 )
 from scripts.game_structure.screen_settings import MANAGER
 from scripts.game_structure.ui_elements import (
-    UIImageButton,
     UITextBoxTweaked,
 )
+from scripts.ui.elements.image_button import UIImageButton
 from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.screens.enums import GameScreen
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
