+++ scripts/ui/windows/select_focus_clans.py
@@ -2,9 +2,7 @@ import pygame
 import pygame_gui
 
 from scripts.game_structure import game
-from scripts.game_structure.ui_elements import (
-    UIImageButton,
-)
+from scripts.ui.elements.image_button import UIImageButton
 from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.screens.enums import GameScreen
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
