+++ scripts/ui/windows/window_base_class.py
@@ -5,7 +5,7 @@ from pygame_gui.elements import UIWindow
 from scripts.game_structure import constants
 from scripts.game_structure.game.settings import game_setting_get
 from scripts.game_structure.screen_settings import MANAGER
-from scripts.game_structure.ui_elements import UIImageButton
+from scripts.ui.elements.image_button import UIImageButton
 from scripts.ui.generate_box import get_box, BoxStyles
 from scripts.ui.scale import ui_scale_value, ui_scale, ui_scale_offset
